Added the simple files at first cut.
This achieves the most important result. That is that unit tests that form part of a ginkgo/gomega test suite can be debugged.
The next step would be to use a package level variable (not ideal but we are constrained by the signatures of test functions) to return the values...